+++ backend/config/routes.php
@@ -151,7 +151,6 @@ return [
     '/api/user/player/with-role/{name}'            => ['GET',    PlayerController::class.'::withRole'],
     '/api/user/player/with-status/{name}'          => ['GET',    PlayerController::class.'::withStatus'],
 
-    '/api/user/settings/system/theme'                       => ['GET',  SettingsController::class.'::theme'],
     '/api/user/settings/system/list'                        => ['GET',  SettingsController::class.'::systemList'],
     '/api/user/settings/system/change/{name}'               => ['PUT',  SettingsController::class.'::systemChange'],
     '/api/user/settings/system/send-invalid-token-mail'     => ['POST', SettingsController::class.'::sendInvalidTokenMail'],