+++ backend/config/security.php
@@ -97,7 +97,6 @@ return [
     '/api/user/player/{id}/set-status/'            => [Role::USER_MANAGER],
     '/api/user/player'                             => [Role::USER_ADMIN],
 
-    '/api/user/settings/system/theme' => [Role::ANONYMOUS, Role::USER],
     '/api/user/settings/system/list'  => [Role::ANONYMOUS, Role::USER],
     '/api/user/settings/system'       => [Role::SETTINGS],