+++ frontend/src/components/NavBar.vue
@@ -127,14 +127,9 @@ export default {
 
     mounted: function() {
         addNavBehaviour();
-        this.selectTheme(this.settings.customization_default_theme);
     },
 
     watch: {
-        settings: function () {
-            this.selectTheme(this.settings.customization_default_theme);
-        },
-
         selectedTheme () {
             const $body = $('body');
             for (const theme of this.themes) {