+++ web/theme.js
@@ -1,0 +1,8 @@
+/**
+ * This file is loaded (blocking) right before the closing "head" tag.
+ */
+
+// Set the default theme. See frontend/src/mixin.js for possible values.
+setTheme('Darkly');
+
+// Add other code below this line - e.g. for user tracking