+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - New: Lock for the watchlist settings: If enabled, only watchlist admins can add and remove alliances 
   and corporations from the list of watched alliances and corporations. Watchlist managers can still edit the
   kick- and allowlist.
-- Change: The watchlist role no longer gives access to all characters from all accounts, but only to accounts that  
+- Change: The watchlist role no longer gives access to all characters from all accounts, but only to accounts that 
   are on a watchlist for which the user has view permission (similar to how it works for the tracking role and member 
   tracking data).
 - Change: Inclusive and more descriptive terminology (changed "red flags" to warnings, blacklist to kicklist or