+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 dd mm yyyy
 
-- **BC break**: Raised minimum required PHP Version to 7.3.0
+- **BC break**: Raised minimum required PHP version to 7.3.0
+- **BC break** (dev only): Raised minimum Node.js version to 10.16.0 (with npm 6.9.0)
 - Fixed error message when adding a new member to a group failed.
 - Changed sleep time for update-player-groups command